move helper file to a helper folder and rename
